+++ src/main/java/com/omar/hex/GameStatus.java
@@ -1,4 +1,4 @@
-package com.omar.model.game;
+package com.omar.hex;
 
 public enum GameStatus {
     ACTIVE, P1WINS, P2WINS;
