+++ src/main/java/com/omar/hex/TurnStatus.java
@@ -1,0 +1,5 @@
+package com.omar.hex;
+
+public enum TurnStatus {
+    P1TURN, P2TURN;
+}
