+++ src/main/java/com/omar/model/Army.java
@@ -1,4 +1,4 @@
-package com.omar.model.faction;
+package com.omar.model;
 
 public class Army {
     private int firepower;
