+++ src/main/java/com/omar/model/Faction.java
@@ -1,4 +1,4 @@
-package com.omar.model.faction;
+package com.omar.model;
 
 
 public class Faction {
