+++ src/main/java/com/omar/model/TileStatus.java
@@ -1,4 +1,4 @@
-package com.omar.model.world;
+package com.omar.model;
 
 public enum TileStatus {
     EMPTY, P1OCCUPIED, P2OCCUPIED;
