--- src/main/java/com/omar/model/game/README.md
@@ -1,1 +1,0 @@
-The code in HexGame and HexMech was copied and refactored from: https://gist.github.com/salamander2/4329783
