--- src/main/java/com/omar/model/game/Turn.java
@@ -1,5 +1,0 @@
-package com.omar.model.game;
-
-public enum Turn {
-    P1TURN, P2TURN;
-}
